Moved class location
